Complete lp_grid_init
Image is still not being output correctly, but at least it's a start!

Hackish way of fixing the issues I was having in grid_init was to add double
parameters for x and y and convert them to the pixel values. Would like to make
a "point" struct to encapsulate x and y and use that instead of pixel. This way
both structs have a coordinate point which are accurate as doubles instead of a
cell having a pixel. This will make the in_circ_funciton easier to write as it
will only compare two points.