Convert to using doubles for location x and y
There was a bunch of refactoring in this commit, adding the point struct and
the lp_grid struct. The point struct allows for a common location type with
x and y double precision floating point values. The grid has also been
encapsulated in its own struct to pass the height and width of the output
grid together in one succinct object. This makes function calls much clearer.
The lp_create_grid function has been added which will initialize the grid
values as well as allocate the space for the grid, making the process
transparent to the user.